Simplify main.dart
